Class Clothing
Changes to constants.

+++ Clothing.java
@@ -18,7 +18,7 @@ public class Clothing extends Item
 
 	/** Current salex tax rate, in decimal format (e.g. 10% = .10)
 	  */
-	private static final int TAX_RATE = .10;
+	private static final float TAX_RATE = .10;
 	
 	/** Current shipping rate.
 	  */
